refactor(desktop): extract library import flow
