+++ apps/desktop/src/features/library-workspace/import/index.ts
@@ -8,3 +8,8 @@ export {
   type ImportSummary,
 } from './LibraryImportModals';
 export { DropImportWarningModal, type DropImportWarningModalProps } from './DropImportWarningModal';
+export { useLibraryImportFlow, type LibraryImportFlow } from './model/useLibraryImportFlow';
+export {
+  LibraryImportFlowModals,
+  type LibraryImportFlowModalsProps,
+} from './ui/LibraryImportFlowModals';
